Refactor to make the project more general over different test types
