+++ cennznet-1.0.0-RC2/docker-compose.yml
@@ -1,0 +1,55 @@
+version: '3'
+services:
+  boot-node-cennz-1_0_0-rc2-1:
+    image: cennznet/cennznet:1.0.0-rc2
+    container_name: boot-node-cennz-1_0_0-rc2-1
+    volumes: ["./data/boot-node-cennz-1_0_0-rc2-1:/mnt/node"]
+    command:
+      - --base-path=/mnt/node
+      - --chain=dev
+      - --validator
+      - --unsafe-ws-external
+      - --ws-port=9944
+      - --unsafe-rpc-external
+      - --rpc-cors=all
+      - --bootnodes=/dns4/boot-node-cennz-1_0_0-rc2-3/tcp/30333/p2p/QmdoyqhDeYScKjNKZTK9Ru3KugwXxiUBaS2k9CESDbn3Be
+      # Some arbitrary static bytes to set the libp2p public key
+      - --node-key=D0FE7E415AA9BAD7605A6841E74C2E0BA72DA1367BBA28AA37DC044669D09858
+
+
+  boot-node-cennz-1_0_0-rc2-2:
+    image: cennznet/cennznet:1.0.0-rc2
+    container_name: boot-node-cennz-1_0_0-rc2-2
+    volumes: ["./data/boot-node-cennz-1_0_0-rc2-2:/mnt/node"]
+    command:
+      - --base-path=/mnt/node
+      - --chain=dev
+      - --validator
+      - --alice
+      - --unsafe-ws-external
+      - --ws-port=9944
+      - --unsafe-rpc-external
+      - --rpc-cors=all
+      - --bootnodes=/dns4/boot-node-cennz-1_0_0-rc2-3/tcp/30333/p2p/QmdoyqhDeYScKjNKZTK9Ru3KugwXxiUBaS2k9CESDbn3Be
+      # Some arbitrary static bytes to set the libp2p public key
+      - --node-key=D0FE7E415AA9BAD7605A6841E74C2E0BA72DA1367BBA28AA37DC044669D09859
+    ports:
+      - "9933:9933"
+      - "9944:9944"
+      - "30333:30333"
+      - "8080:8000"
+
+  boot-node-cennz-1_0_0-rc2-3:
+    image: cennznet/cennznet:1.0.0-rc2
+    container_name: boot-node-cennz-1_0_0-rc2-3
+    volumes: ["./data/boot-node-cennz-1_0_0-rc2-3:/mnt/node"]
+    command:
+      - --base-path=/mnt/node
+      - --chain=dev
+      - --validator
+      - --unsafe-ws-external
+      - --ws-port=9944
+      - --unsafe-rpc-external
+      - --rpc-cors=all
+      # Some arbitrary static bytes to set the libp2p public key
+      - --node-key=D0FE7E415AA9BAD7605A6841E74C2E0BA72DA1367BBA28AA37DC044669D09860
