+++ plug_1-0-0_RC1/docker-compose.yml
@@ -1,0 +1,78 @@
+version: '3'
+services:
+  node-1:
+    image: plugnet/plugblockchain:1.0.0-rc1
+    container_name: plug-node-1.0.0-rc1-1
+    volumes: ["./data/plug_node-1:/mnt/node"]
+    command:
+      - /usr/local/bin/substrate
+      - --base-path=/mnt/node
+      - --chain=dev
+      - --bob
+      - --validator
+      - --ws-external
+      - --ws-port=9944
+      - --rpc-external
+      - --rpc-cors=all
+      # Some arbitrary static bytes to set the libp2p public key
+      - --node-key=D0FE7E415AA9BAD7605A6841E74C2E0BA72DA1367BBA28AA37DC044669D09858
+    ports:
+      - "9933:9933"
+      - "9944:9944"
+      - "30333:30333"
+      - "8080:8000"
+  node-2:
+    image: plugnet/plugblockchain:1.0.0-rc1
+    container_name: plug-node-1.0.0-rc1-2
+    volumes: ["./data/plug_node-2:/mnt/node"]
+    command:
+      - /usr/local/bin/substrate
+      - --base-path=/mnt/node
+      - --chain=dev
+      - --bootnodes=/dns4/plug-node-1.0.0-rc1-1/tcp/30333/p2p/QmQmJTi2eDg53Ws3N5XaixMmiGCmhNNJX7bPBMirnD2JEs
+      - --alice
+      - --validator
+      - --ws-external
+      - --rpc-cors=all
+
+  node-3:
+    image: plugnet/plugblockchain:1.0.0-rc1
+    container_name: plug-node-1.0.0-rc1-3
+    volumes: ["./data/plug_node-3:/mnt/node"]
+    command:
+      - /usr/local/bin/substrate
+      - --base-path=/mnt/node
+      - --chain=dev
+      - --bootnodes=/dns4/plug-node-1.0.0-rc1-1/tcp/30333/p2p/QmQmJTi2eDg53Ws3N5XaixMmiGCmhNNJX7bPBMirnD2JEs
+      - --charlie
+      - --validator
+      - --ws-external
+      - --rpc-cors=all
+
+  node-4:
+    image: plugnet/plugblockchain:1.0.0-rc1
+    container_name: plug-node-1.0.0-rc1-4
+    volumes: ["./data/plug_node-4:/mnt/node"]
+    command:
+      - /usr/local/bin/substrate
+      - --base-path=/mnt/node
+      - --chain=dev
+      - --bootnodes=/dns4/plug-node-1.0.0-rc1-1/tcp/30333/p2p/QmQmJTi2eDg53Ws3N5XaixMmiGCmhNNJX7bPBMirnD2JEs
+      - --dave
+      - --validator
+      - --ws-external
+      - --rpc-cors=all
+
+  node-5:
+    image: plugnet/plugblockchain:1.0.0-rc1
+    container_name: plug-node-1.0.0-rc1-5
+    volumes: ["./data/plug_node-5:/mnt/node"]
+    command:
+      - /usr/local/bin/substrate
+      - --base-path=/mnt/node
+      - --chain=dev
+      - --bootnodes=/dns4/plug-node-1.0.0-rc1-1/tcp/30333/p2p/QmQmJTi2eDg53Ws3N5XaixMmiGCmhNNJX7bPBMirnD2JEs
+      - --eve
+      - --validator
+      - --ws-external
+      - --rpc-cors=all
